// client/src/services/types.ts
export type GetNotesData = {
  id: number;
  title: string;
  body: string;
  createdAt: string;
  updatedAt: string;
};

export type AddNoteData = {
  id: number;
  title: string;
  body: string;
  createdAt: string;
  updatedAt: string;
};

export type UpdateNoteData = {
  id: number;
  title: string;
  body: string;
  createdAt: string;
  updatedAt: string;
};